+++ pages/formula.php
@@ -1,7 +1,10 @@
 <?php 
-if (!defined('pvault_panel')){ die('Not Found');}  
-$f_name =  mysqli_real_escape_string($conn, $_GET['name']);
-$fid = base64_encode($f_name);
+if (!defined('pvault_panel')){ die('Not Found');}
+require_once(__ROOT__.'/func/arrFilter.php');
+require(__ROOT__.'/func/get_formula_notes.php');
+$fid = mysqli_real_escape_string($conn, $_GET['name']);
+$f_name =  base64_decode($fid);
+
 
 if(mysqli_num_rows(mysqli_query($conn, "SELECT id FROM formulasMetaData WHERE fid = '$fid'")) == FALSE){
 	echo 'Formula doesn\'t exist';
@@ -11,6 +14,35 @@ if(mysqli_num_rows(mysqli_query($conn, "SELECT fid FROM formulas WHERE fid = '$f
 	$legend = 1;
 }
 $meta = mysqli_fetch_array(mysqli_query($conn, "SELECT id,image FROM formulasMetaData WHERE fid = '$fid'"));
+
+$formula_q = mysqli_query($conn, "SELECT ingredient FROM formulas WHERE fid = '$fid'");
+	while ($formula = mysqli_fetch_array($formula_q)){
+		$form[] = $formula;
+	}
+	
+						
+	foreach ($form as $formula){
+		$top_ing = mysqli_fetch_array(mysqli_query($conn, "SELECT name AS ing,category FROM ingredients WHERE name = '".$formula['ingredient']."' AND profile = 'Top' AND category IS NOT NULL"));
+		$heart_ing = mysqli_fetch_array(mysqli_query($conn, "SELECT name AS ing,category FROM ingredients WHERE name = '".$formula['ingredient']."' AND profile = 'Heart' AND category IS NOT NULL"));
+		$base_ing = mysqli_fetch_array(mysqli_query($conn, "SELECT name AS ing,category FROM ingredients WHERE name = '".$formula['ingredient']."' AND profile = 'Base' AND category IS NOT NULL"));
+	
+		$top_cat = mysqli_fetch_array(mysqli_query($conn, "SELECT name FROM ingCategory WHERE id = '".$top_ing['category']."' AND image IS NOT NULL"));
+		$heart_cat = mysqli_fetch_array(mysqli_query($conn, "SELECT name FROM ingCategory WHERE id = '".$heart_ing['category']."' AND image IS NOT NULL"));
+		$base_cat = mysqli_fetch_array(mysqli_query($conn, "SELECT  name FROM ingCategory WHERE id = '".$base_ing['category']."' AND image IS NOT NULL"));
+	
+		$top[] = array_merge($top_cat,$top_ing);
+		$heart[] = array_merge($heart_cat,$heart_ing);
+		$base[] = array_merge($base_cat,$base_ing);
+
+	}
+$top_cat = array_filter($top);
+$heart_cat = array_filter($heart);
+$base_cat = array_filter($base);
+
+
+$top_ex = get_formula_excludes($conn, $fid, 'top');
+$heart_ex = get_formula_excludes($conn, $fid, 'heart');
+$base_ex = get_formula_excludes($conn, $fid, 'base');
 ?>
 <style>
 .mfp-iframe-holder .mfp-content {
@@ -27,7 +59,7 @@ $meta = mysqli_fetch_array(mysqli_query($conn, "SELECT id,image FROM formulasMet
           <div class="card shadow mb-4">
             <div class="card-header py-3"> 
 			  <?php if($meta['image']){?><div class="img-formula"><img class="img-perfume" src="<?php echo $meta['image']; ?>"/></div><?php } ?>
-              <h2 class="m-0 font-weight-bold text-primary"><a href="?do=Formula&name=<?=base64_decode($fid)?>"><?=base64_decode($fid)?></a></h2>
+              <h2 class="m-0 font-weight-bold text-primary"><a href="?do=Formula&name=<?=$fid?>"><?=$f_name?></a></h2>
               <h5 class="m-1 text-primary"><a href="pages/getFormMeta.php?id=<?php echo $meta['id'];?>" class="popup-link">Details</a></h5>
             </div>
         <!-- Nav tabs -->
@@ -114,8 +146,9 @@ $meta = mysqli_fetch_array(mysqli_query($conn, "SELECT id,image FROM formulasMet
 		        <div id="fetch_summary"><div class="loader"></div></div>
                 <?php if($legend){ ?>
                 <div id="share">
-               	  <p>To include this page in your web site, copy this line and paste it into your html code</p>
-               	  <p><pre>&lt;iframe src=&quot;<?=$_SERVER['REQUEST_SCHEME']?>://<?=$_SERVER['SERVER_NAME']?>/pages/viewSummary.php?id=<?=$fid?>&quot; title=&quot;<?=base64_decode($fid)?>&quot;&gt;&lt;/iframe&gt;</pre></p>
+               	  <p><a href="#" data-toggle="modal" data-target="#conf_view">Configure view</a></p>
+               	  <p>To include this page in your web site, copy this line and paste it into your html code:</p>
+           	    <p><pre>&lt;iframe src=&quot;<?=$_SERVER['REQUEST_SCHEME']?>://<?=$_SERVER['SERVER_NAME']?>/pages/viewSummary.php?id=<?=$fid?>&quot; title=&quot;<?=$f_name?>&quot;&gt;&lt;/iframe&gt;</pre></p>
                 	<p>For documentation and parameterisation please refer to: <a href="https://www.jbparfum.com/knowledge-base/share-formula-notes/" target="_blank">https://www.jbparfum.com/knowledge-base/share-formula-notes/</a></p>
                 </div>
                 <?php } ?>
@@ -127,6 +160,108 @@ $meta = mysqli_fetch_array(mysqli_query($conn, "SELECT id,image FROM formulasMet
      </div><!--tabs-->
    </div>
   </div>
+  
+<!--Configure View-->
+
+<div class="modal fade" id="conf_view" tabindex="-1" role="dialog" aria-labelledby="conf_view" aria-hidden="true">
+  <div class="modal-dialog modal-conf-view" role="document">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h5 class="modal-title" id="conf_view">Choose which notes will be displayed</h5>
+        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+          <span aria-hidden="true">&times;</span>
+        </button>
+      </div>
+      <div class="modal-body">
+   	    <div id="confViewMsg"></div>
+          <form action="javascript:update_view()" id="form1">   
+           <div class="conf_tbl">
+            <table width="100%" border="0">
+              <tr>
+                <td colspan="2"><strong>Top notes</strong><hr /></td>
+              </tr>
+              <?php foreach ($top_cat as $x){
+						if (!is_numeric(array_search($x['name'],$top_ex ))){
+				?>
+              <tr>
+				<td width="54%" height="29" ex_top_ing_name="<?=$x['name']?>"><?=$x['name']?></td>
+                <td width="46%"><input name="ex_top_ing" class="ex_ing" type="checkbox" id="<?=str_replace(' ', '_',$x['ing'])?>" value="<?=str_replace(' ', '_',$x['ing'])?>" checked="checked" /></td>
+              </tr>
+              <?php }else{ ?>
+			  <tr>
+				<td width="54%" ex_top_ing_name="<?=$x['name']?>"><?=$x['name']?></td>
+                <td width="46%"><input name="ex_top_ing" class="ex_ing" type="checkbox" id="<?=str_replace(' ', '_',$x['ing'])?>" value="<?=str_replace(' ', '_',$x['ing'])?>" /></td>
+              </tr>
+			 <?php 
+			 	}
+			  }
+			  ?>
+             </table>
+            </div>
+             
+           <div class="conf_tbl">
+            <table width="100%" border="0">
+              <tr>
+                <td colspan="2"><p><strong>Heart notes</strong></p><hr /></td>
+              </tr>
+              <?php foreach ($heart_cat as $x){
+						if (!is_numeric(array_search($x['name'],$heart_ex ))){
+			   ?>
+              <tr>
+				<td width="40%" height="29"><?=$x['name']?></td>
+                <td width="51%"><input name="ex_heart_ing" class="ex_ing" type="checkbox" id="<?=str_replace(' ', '_',$x['ing'])?>" value="<?=str_replace(' ', '_',$x['ing'])?>" checked="checked" /></td>
+              </tr>
+              <?php }else{ ?>
+              <tr>
+				<td><?=$x['name']?></td>
+                <td width="51%"><input name="ex_heart_ing" class="ex_ing" type="checkbox" id="<?=str_replace(' ', '_',$x['ing'])?>" value="<?=str_replace(' ', '_',$x['ing'])?>" /></td>
+              </tr>
+              <?php 
+			 	}
+			  }
+			  ?>
+             </table> 
+            </div>
+
+            <div class="conf_tbl">
+             <table width="100%" border="0">
+              <tr>
+                <td colspan="2"><p><strong>Base notes</strong></p><hr /></td>
+              </tr>
+              <?php foreach ($base_cat as $x){
+						if (!is_numeric(array_search($x['name'],$base_ex ))){
+			  ?>
+              <tr>
+				<td width="40%" height="29"><?=$x['name']?></td>
+                <td width="60%"><input name="ex_base_ing" class="ex_ing" type="checkbox" id="<?=str_replace(' ', '_',$x['ing'])?>" value="<?=str_replace(' ', '_',$x['ing'])?>" checked="checked" /></td>
+              </tr>
+             <?php }else{ ?>
+              <tr>
+				<td><?=$x['name']?></td>
+                <td width="60%"><input name="ex_base_ing" class="ex_ing" type="checkbox" id="<?=str_replace(' ', '_',$x['ing'])?>" value="<?=str_replace(' ', '_',$x['ing'])?>" /></td>
+              </tr>
+              <?php 
+			 	}
+			  }
+			  ?>
+            </table>
+            </div>
+            <table width="100%" border="0">
+              <tr>
+                <td>            
+  					<div class="modal-footer">
+     	  				<button type="button" class="btn btn-secondary" data-dismiss="modal">Cancel</button>
+	 	  				<input type="submit" name="button" class="btn btn-primary" id="btnUpdate" value="Save">
+   		  			</div>
+           			</td>
+    			</tr>
+			</table>
+          </form>
+    </div>
+  </div>
+  </div>
+</div>
+
 <script type="text/javascript" language="javascript" >
 //$(document).ready(function(){
  //UPDATE PURITY
@@ -288,4 +423,26 @@ $.ajax({
 
 fetch_summary();
 
+function update_view(){
+	
+	$('.ex_ing').each(function(){
+		$.ajax({ 
+			url: 'pages/manageFormula.php', 
+			type: 'get',
+			data: {
+				fid: "<?=urlencode($fid)?>",
+				manage_view: '1',
+				ex_status: $("#" + $(this).val()).is(':checked'),
+				ex_ing: $(this).val()
+				},
+			dataType: 'html',
+				success: function (data) {
+					$('#confViewMsg').html(data);
+					fetch_summary();
+
+				}
+		});
+	});
+
+}
 </script>